[FEAT] Shadow map support (#17)
* group render and render override

* render pass support

* engine material and shadow support

* fix shadow shader and enable shadows

* improve shadows

* shadow PCF

* VSM shadows instead of PCF

* fix light dir

* EVSM shadow complete

* LWGL update

* Switch to VSM and hw accelerated gauss blur

the linear sampling blur is same as 5tap gauss with just 3 sample values, utilizing linear filtering of the texture

* Update RenderContext.h

* Update lwgl

* Update seuil

* missing clang include

+++ engine/src/physics/AABBCollider.cpp
@@ -26,8 +26,8 @@ AABBCollider::AABBCollider(std::shared_ptr<AABB> aabb, float stepHeight)
         throw std::runtime_error("AABBCollider: AABB height must be greater than 0.");
 
     AABB bb(m_aabb->min - s_checkBox, m_aabb->max + s_checkBox);
-    int volume = glm::compMul(bb.max - bb.min);
-    m_aabbCache.reserve(volume - (m_aabb->max.y - m_aabb->min.y));
+    int volume = static_cast<int>(glm::compMul(bb.max - bb.min));
+    m_aabbCache.reserve(static_cast<size_t>(volume - (m_aabb->max.y - m_aabb->min.y)));
 }
 
 void AABBCollider::setWorld(std::shared_ptr<World> world) {
